상품 등록, 수정, 삭제

+++ src/main/java/com/vision/shoppingmall/category/service/CategoryService.java
@@ -1,6 +1,7 @@
 package com.vision.shoppingmall.category.service;
 
 import com.vision.shoppingmall.category.model.entity.Category;
+import com.vision.shoppingmall.category.model.exception.CategoryHasProductException;
 import com.vision.shoppingmall.category.model.exception.CategoryNameDuplicationException;
 import com.vision.shoppingmall.category.model.exception.CategoryNotFoundException;
 import com.vision.shoppingmall.category.model.request.CategoryUpdateRequest;
@@ -9,16 +10,23 @@ import com.vision.shoppingmall.category.model.response.CategoryCreateResponse;
 import com.vision.shoppingmall.category.model.response.CategoryListResponse;
 import com.vision.shoppingmall.category.model.response.CategoryResponse;
 import com.vision.shoppingmall.category.repository.CategoryRepository;
+import com.vision.shoppingmall.product.model.entity.Product;
+import com.vision.shoppingmall.product.model.entity.ProductStatus;
+import com.vision.shoppingmall.product.service.ProductService;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final ProductService productService;
 
     public CategoryCreateResponse create(CreateCategoryRequest request) {
         //1. 카테고리 이름 중복 검사
@@ -41,7 +49,8 @@ public class CategoryService {
                 .map( category ->
                         new CategoryListResponse(
                                 category.getId(),
-                                category.getCategoryName())
+                                category.getCategoryName(),
+                        category.getProducts().stream().filter(product -> product.getProductStatus() == ProductStatus.ACTIVE).count())
                 );
     }
 
@@ -66,12 +75,30 @@ public class CategoryService {
         categoryRepository.save(category);
     }
 
+    @Transactional
     public void delete(Long id) {
         //1. 카테고리가 존재 하는지 확인
         Category category = categoryRepository.findById(id)
                 .orElseThrow(CategoryNotFoundException::new);
+        boolean hasActiveProducts =  category.getProducts().stream().anyMatch(product -> product.getProductStatus() == ProductStatus.ACTIVE);
+        if(hasActiveProducts)
+            throw new CategoryHasProductException();
+
+        List<Product> inActiveProducts = category.getProducts().stream().filter(product -> product.getProductStatus() == ProductStatus.INACTIVE).toList();
+
+        productService.removeCategory(inActiveProducts);
         //2. 카테고리 삭제
         categoryRepository.delete(category);
+    }
+
+    public List<CategoryResponse> getAllCategories() {
+        List<Category> categories = categoryRepository.findAll();
+
+        return categories.stream().map(category ->
+                new CategoryResponse(
+                category.getId(),
+                category.getCategoryName()
+        )).toList();
     }
 
 }
